+++ packages/envapt/src/common.ts
@@ -1,6 +1,7 @@
 export { Environment } from './engine/Envapter';
 export { Converters, isArrayOf } from './converters';
 export { ManualEnvSource } from './sources/ManualEnvSource';
+export { PortableSource } from './sources/PortableSource';
 export { WorkerEnvSource } from './sources/WorkerEnvSource';
 export type { ArrayElement, ArrayOf, ConverterToken, CustomElementConverter } from './converters';
 export type { DebugLevel } from './infra/Debug';